+++ src/main/java/fll/web/admin/CommitAwardGroups.java
@@ -25,7 +25,6 @@ import fll.db.Queries;
 import fll.web.ApplicationAttributes;
 import fll.web.BaseFLLServlet;
 import fll.web.SessionAttributes;
-import net.mtu.eggplant.util.sql.SQLFunctions;
 
 /**
  * Commit the changes from edit_event_division.jsp.
@@ -37,6 +36,10 @@ public class CommitAwardGroups extends BaseFLLServlet {
 
   /**
    * Put the needed variables into the page context for the page to use.
+   * 
+   * @param application used for application variables
+   * @param session used for session variables
+   * @param pageContext used to populate page variables
    */
   public static void populateContext(final ServletContext application,
                                      final HttpSession session,
@@ -44,10 +47,7 @@ public class CommitAwardGroups extends BaseFLLServlet {
     final StringBuilder message = new StringBuilder();
 
     final DataSource datasource = ApplicationAttributes.getDataSource(application);
-    Connection connection = null;
-    try {
-      connection = datasource.getConnection();
-
+    try (Connection connection = datasource.getConnection()) {
       final int currentTournamentID = Queries.getCurrentTournament(connection);
 
       pageContext.setAttribute("divisions", Queries.getAwardGroups(connection, currentTournamentID));
@@ -61,8 +61,6 @@ public class CommitAwardGroups extends BaseFLLServlet {
           + "</p>");
       LOGGER.error(sqle, sqle);
       throw new RuntimeException("Error saving team data into the database", sqle);
-    } finally {
-      SQLFunctions.close(connection);
     }
 
     SessionAttributes.appendToMessage(session, message.toString());
@@ -79,9 +77,7 @@ public class CommitAwardGroups extends BaseFLLServlet {
     final StringBuilder message = new StringBuilder();
 
     final DataSource datasource = ApplicationAttributes.getDataSource(application);
-    Connection connection = null;
-    try {
-      connection = datasource.getConnection();
+    try (Connection connection = datasource.getConnection()) {
 
       final int currentTournamentID = Queries.getCurrentTournament(connection);
 
@@ -96,8 +92,8 @@ public class CommitAwardGroups extends BaseFLLServlet {
               && !newAwardGroup.isEmpty()) {
             if ("text".equals(newAwardGroup)) {
               // get from text box
-              newAwardGroup = request.getParameter("text_"
-                  + teamNumberStr);
+              newAwardGroup = request.getParameter(teamNumberStr
+                  + "_text");
             }
             if (null != newAwardGroup
                 && !newAwardGroup.trim().isEmpty()) {
@@ -120,8 +116,6 @@ public class CommitAwardGroups extends BaseFLLServlet {
           + "</p>");
       LOGGER.error(sqle, sqle);
       throw new RuntimeException("Error saving award group data into the database", sqle);
-    } finally {
-      SQLFunctions.close(connection);
     }
 
     SessionAttributes.appendToMessage(session, message.toString());