+++ Changes.md
@@ -1,3 +1,5 @@
+* #818 - Fix setting of custom judging and award groups
+
 Release 14.4.0
 ==============
 * #813 - Allow the user to specify the order that the awards groups appear in the awards report